Changes in INITDB

+++ openwords_android_client/src/com/openwords/model/InitDatabase.java
@@ -3,10 +3,12 @@ package com.openwords.model;
 import com.openwords.DAO.PlateDbHelper;
 import com.openwords.DAO.UserPerfDbHelper;
 import com.openwords.DAO.UserWordsDbHelper;
+import com.openwords.util.preference.OpenwordsSharedPreferences;
 
 import android.content.Context;
 
 public class InitDatabase {
+	public static String url_get_user_perf_from_server = "";
 	public Context context;
 	public PlateDbHelper plateDb;
 	public UserPerfDbHelper userPerfDb;
@@ -24,21 +26,27 @@ public class InitDatabase {
 		this.userWordsDb = new UserWordsDbHelper(this.context);
 	}
 	
-	public void RefreshTables(String url_get_from_server)
+	public void RefreshTables()
 	{
-		this.plateDb.deleteAll();
-		this.userPerfDb.deleteAllUserPerf();
-		this.userWordsDb.deleteAllUserWords();
-		this.userWordsDb.deleteAllTranscriptions();
+		int userId = OpenwordsSharedPreferences.getUserInfo().getUserId();
+		int userIdFromDb = this.userPerfDb.getUserFromUserPerf();
 		
-		//-----read from server-----
-		try
+		if(userId != userIdFromDb)
 		{
-			//------------ for User Perf-------
-			//------------ for User words and TRanscriptions-------
-		}catch(Exception e)
-		{
-			e.printStackTrace();
+			this.plateDb.deleteAll();
+			this.userPerfDb.deleteAllUserPerf();
+			this.userWordsDb.deleteAllUserWords();
+			this.userWordsDb.deleteAllTranscriptions();
+			
+			//-----read from server-----
+			try
+			{
+				//------------ for User Perf-------
+				//------------ for User words and TRanscriptions-------
+			}catch(Exception e)
+			{
+				e.printStackTrace();
+			}
 		}
 	}
